Replace a bunch of size_t with unsigned
I've apparently gotten into the wrong mindset about size_t vs
unsigned.  It's easy to let size_t proliferate throughout your
code base once you start matching types with all of the standard
library arguments and thinking of this as the "natural" type for
anything that counts or sizes things.  It seems to have a certain
cleanliness advantage in that it won't wrap in the real world,
too.

However, it's commonly wasteful (e.g. x86_64) when the sizes are
known to be reasonably-limited, and better to stick with unsigned
in the general case.

We've kept size_t in places where it clearly is a good idea: when
it's directly storing a size_t return value from a function before
checking and converting, for our own library functions related to
raw allocation and file-mapping, and for file data offsets in
readers of potentially-large files (like the zonefile scanner).

+++ include/gdnsd/dname.h
@@ -207,7 +207,7 @@ gdnsd_dname_status_t gdnsd_dname_status(const uint8_t* dname);
 // non-NULL pointer, but a buflen of zero is legal (which will always return
 // DNAME_INVALID).
 F_NONNULL F_PURE F_UNUSED
-static gdnsd_dname_status_t gdnsd_dname_status_buflen(const uint8_t* dname, size_t buflen)
+static gdnsd_dname_status_t gdnsd_dname_status_buflen(const uint8_t* dname, unsigned buflen)
 {
     if (!buflen || ((*dname) + 1U) > buflen)
         return DNAME_INVALID;
@@ -264,7 +264,7 @@ static void gdnsd_dname_copy(uint8_t* dest, const uint8_t* source)
 F_MALLOC F_NONNULL F_UNUSED F_RETNN
 static uint8_t* gdnsd_dname_dup(const uint8_t* dname)
 {
-    const size_t sz = *dname + 1U;
+    const unsigned sz = *dname + 1U;
     uint8_t* out = xmalloc(sz);
     memcpy(out, dname, sz);
     return out;
